Attempt to limit modal analysis to just Velbert area
This is an attempt to geometrically limit the analysis. It unfortunately does not work because only basing it off of person id means that if they return back into the boundary, they will receive an empty mode "" in the activity end event, but then it won't get replaced in the departure event.