Fix the modal share analysis to correctly limit modes to Velbert area
A new hash map was created that stores the coordinates of every activity start for each person id. The departure events then looks at the most recent activity start location, and if it is in the boundary, it adds the mode. If it is outside the boundary, it does not add the mode.